// apps/ui/src/lib/api-client/models/IdentityProviderResponse.ts
export interface IdentityProviderResponse {
    /**
     * Provider alias
     * @type {string}
     * @memberof IdentityProviderResponse
     */
    alias?: string;
    /**
     * Provider type
     * @type {string}
     * @memberof IdentityProviderResponse
     */
    providerId?: string;
    /**
     * Display name
     * @type {string}
     * @memberof IdentityProviderResponse
     */
    displayName?: string;
    /**
     * Whether provider is enabled
     * @type {boolean}
     * @memberof IdentityProviderResponse
     */
    enabled?: boolean;
    /**
     * @type {{ [key: string]: any }}
     * @memberof IdentityProviderResponse
     */
    config?: { [key: string]: any };
    /**
     * @type {boolean}
     * @memberof IdentityProviderResponse
     */
    addReadTokenRoleOnCreate?: boolean;
    /**
     * @type {string}
     * @memberof IdentityProviderResponse
     */
    firstBrokerLoginFlowAlias?: string;
    /**
     * @type {string}
     * @memberof IdentityProviderResponse
     */
    internalId?: string;
    /**
     * @type {boolean}
     * @memberof IdentityProviderResponse
     */
    linkOnly?: boolean;
    /**
     * @type {boolean}
     * @memberof IdentityProviderResponse
     */
    hideOnLogin?: boolean;
    /**
     * @type {string}
     * @memberof IdentityProviderResponse
     */
    postBrokerLoginFlowAlias?: string;
    /**
     * @type {boolean}
     * @memberof IdentityProviderResponse
     */
    storeToken?: boolean;
    /**
     * @type {boolean}
     * @memberof IdentityProviderResponse
     */
    trustEmail?: boolean;
    /**
     * @type {string}
     * @memberof IdentityProviderResponse
     */
    organizationId?: string;
}